Split view selection into Selection and DocumentSelection.
